TGB/requirements.txt
geopandas==0.14.0
matplotlib==3.8.2
numpy==1.26.4
open3d==0.16.1
opencv_contrib_python==4.8.0.74
Pillow==10.2.0
scipy==1.12.0
Shapely==2.0.3
